Change agent's action strategy on close request
Previously, the agent used to close the dialog as soon as the user
requested a dialog termination. This hurts user simulation since the
user simulator learns that taking the "close" action at any stage leads
to immediate dialog completion.

Now, the agent closes the dialog only when it has all the information,
ignoring user's "close" requests in other states.